Update mongodb versions - small tweaks in app - still can't run
mongodb connection is failing because the privileges aren't right or
something.. idk.
am confused about it haha

_
my understanding so far is that the call to `app.use( session( ...` in
server.js is the issue, and somehow the uri being used there isn't
valid? (even though it's the same uri??)
rather - perhaps i need to grant my mongodb user permission for creating
that store?